// src/types/Connector.ts
export type Field = {
  key: string;
  label?: string;
  helpText?: string;
  type?: string;
  /*| "string"
      | "text"
      | "integer"
      | "number"
      | "boolean"
      | "datetime"
      | "file"
      | "password"
      | "copy"
      | "code";*/
  required?: boolean;
  placeholder?: string;
  default?: string;
  choices?: FieldChoice[];
  list?: boolean;
  children?: Field[];
  dict?: boolean;
  computed?: boolean;
  updateFieldDefinition?: boolean;
  inputFormat?: string;
  validation?: any;
  readonly?: boolean;
};

export type FieldChoice =
  | string
  | {
      value: string;
      label: string;
      sample: string;
      icon?: string;
    };